Initial upload
generate_emoji_list.pl does a nice job converting the official unicode.org to an easily digestible list.
emojipick.sh can use either dmenu or fsf to browse the list and send the chosen one to the clipboard. Unfortunately, both options are glitchy and randomly manage to render the glyph or not. More unfortunately, when rendering it badly, it's also copies that garbage to the clipboard.